video: Don't double-convert when parsing `TCOEF`

+++ h263/src/parser/block.rs
@@ -733,13 +733,13 @@ where
                         tcoef.push(TCoefficient {
                             is_short: true,
                             run,
-                            level: level as i8 as i16,
+                            level: level as i16,
                         })
                     } else {
                         tcoef.push(TCoefficient {
                             is_short: true,
                             run,
-                            level: -(level as i8 as i16),
+                            level: -(level as i16),
                         })
                     }
